updated styling and layouts
