+++ src/app/not-found.tsx
@@ -1,0 +1,5 @@
+import { ParkingPage } from "@/components/common/parking";
+
+export default function NotFound() {
+  return <ParkingPage />;
+}
